Set the size of the background to zero
If the minimum object size is greater than the background pixel volume, then the background will also be included as an "object to merge". When calculating relative surface area, this would cause a divide by zero error to occur. Since the background isn't technically an object we want to merge, just set it to zero so it isn't considered.