Create RunImgeJScript Module (#120)
* Create initial imagej script module

* RunImageJScript: auto-parse script params

* RunImageJScript: only parse valid scripts

* Create pyimagej daemon to process commands

* Clear script params on script change

* RunImageJScript: add module validation

* Initial support for converting script i/o types

* Enable script execution

* Refactor to use dictionaries

Relying on calling Queue.get() the appropriate number of times feels too
fragile with the possibility of stalling indefinitely. Using
dictionaries allows us to be a bit more explicit in how to communicate
between pyimagej and CP

* Add progress indicator when parsing params

Initializing the pyimagej daemon is slow, so this way we give the user
an indication that something is happening at least.

* runimagejscript: fix setting construction

* runimagejscript: move all UI interaction

This needs to happen on the same thread that is launched when a button
is clicked.

* Add input validation

* Add module doc

* Fix unknown command note

* Preprocess images to convert to datasets

* module help text

* Factor out settings classes

* Fix deadlock when script doesn't exist

* saving settings

* remover button

* saving and loading parsed settings is done

* Remove automodule import of java classes

For some reason this stopped working, but it seems cleaner to use the
python objects anyway and let jpype autoconvert.

* Clear script parameters when changing script

Instead of the value callback, which would be triggered when loading
values from saved workflows, we now just clear the previous script
params when the selection button is used.

* Code cleanup

* Do not discard settings collections

These are necessary for tracking counts so re-initializing them will
cut the connection with the hidden setting.

* Clear all settings when parsing new script

* Add output settings

* fix i/o state preservation

* Create specialized output settings for renaming

* Consolidate hidden + visible settings

For the remover to operate correctly it has to remove the hidden
settings as well as the visible, so they all need to go into a single
unified group.

* Extract static method

* Convert returned java types to python

* Formatting

* Use ImageName instead of Text for outputs

* Only commit to OUTPUT images

* Show input/output display

* RunImageJScript: fix validation

ValidationError needs a setting, and needs to be raised if the file no
longer exists.

* Improve check for script existence

Stops a bug where if the script is empty but the directory is valid, the
test would go on.

* Empty tests

* test imagej process

* test parse parameters

* renamed test file

* cleanup test

* test copy image

* Add Fiji executable path as setting

* Add a warning if user did not specify a valid Fiji executable path

* Update requirements.txt

RunImageJScript requires pyimagej, which is incompatible with the plain
"imagej" dependency.

* Fix parameter loading when no parsed settings

* docs

* Refactor initialization options

Users can now select from the 3 primary methods of initialization
ImageJ: a local installation, an endpoint specification, or the latest
Fiji.

Users are warned when selecting an option that may download files to
their local machine.

Users are now properly warned when initialization fails as well.

Additionally, the behavior of Directory/Filename settings differs
slightly between Mac and Windows. Selecting local apps should now be a
consistent experience on both platforms.

* Make ImageJ process global

This ensures there is only one process and it can be shared between
module instances. This prevents multiple processes from lingering when
more modules are created (or reloaded).

* Typo fix

* Remove unnecessary pass statements

* moved old conftest to CP3 folder

* updated conftest to use cellprofiler_core instead of cellprofiler

* remove unused tests

Co-authored-by: alucas <alicelucas93@gmail.com>