Add support for multi-setting script Parameters
By mapping File > Filename then the files have to be in the same working
directory as cellprofiler. Instead, files in CellProfiler are handled
with multiple settings (a Directory + Filename).

We now properly map File script parameters to Directory + Filename
settings in CP. The setting logic should now be flexible enough to
handle future multi-setting mappings if needed.

NB: script parameters in a previously saved project file are
incompatible with this change and will need to be re-parsed upon
loading.