CP-IJ: Multiprocessing refactor (#144)
* CP-IJ: Factor out multiprocessing to subclass

The ImageJ daemon doesn't need to know anything about CellProfiler and
in some cases it seems beneficial to have the multiprocess code isolated
(e.g. to define a main method which may be needed for some
 multiprocessing contexts).

* CP-IJ bridge: add __main__

This can be necessary when running subprocesses to give the new
interpreter a starting point.

* CP-IJ: Refactor to server-based model

This works around the blocked PIPE communication between subprocesses.

* Actualize virtual ImagePlus

* CP-IJ: add documentation

* Remove unused code

* Switch to standard Queue

* Ensure IJ server starts when needed

There are two places we need to check if the server should start:
- prepare_settings covers when loading a workflow
- get_parameters_from_script covers when starting with a fresh module

* Simplify IJ server shutdown

* Add freeze_support

* Rename server_check method

* Add method to wait for server startup

* Add more docstrings

* Update remaining work

* CP-IJ: synchronize communication with server

* CP-IJ: restore init string functionality

* CP-IJ: remove unused KEY_ERROR

* CP-IJ: fix typo

* CP-IJ: remove print statement

* CP-IJ: add manual sleep in addition to timeout

On OSX we identified scenarios where the socket.connect returns
immediately on failure, even when a timeout is given. So we now manually
sleep between attempts as well.

* Update README.md

Add RunImageJScript notes

* CP-IJ: handle scenario where imagej startup fails

* CP-IJ: remove cached server proxies

Too big of a liability for server skew, e.g. if the server is restarted.

* CP-IJ: add type hints to proxy accessors

* CP-IJ: link image subplots